+++ src/main/java/id/kedukasi/core/repository/wilayah/ProvinsiRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import id.kedukasi.core.models.wilayah.Provinsi;
 
 @Repository
-public interface ProvinsiRepository extends JpaRepository<Provinsi, Integer>{
+public interface ProvinsiRepository extends JpaRepository<Provinsi, Long>{
     
 }
