+++ src/main/java/id/kedukasi/core/service/MentorService.java
@@ -11,8 +11,8 @@ public interface MentorService {
   
     ResponseEntity<?> updateMentor(Long id, String nama_mentor, String kode,MultipartFile foto, String no_ktp,
     String no_telepon, String status, String class_name, String pendidikan_univ,
-    String pendidikan_jurusan, Date tgl_start, Date tgl_stop,  String alamat_rumah, MultipartFile cv, Integer provinsi, Integer kota,
-    Integer kecamatan, Integer kelurahan);
+    String pendidikan_jurusan, Date tgl_start, Date tgl_stop,  String alamat_rumah, MultipartFile cv, Long provinsi, Long kota,
+    Long kecamatan, Long kelurahan);
     
     ResponseEntity<?> deleteMentor(boolean banned, long id, String uri);
